pyproject.toml
[tool.poetry]
name = "lunar-encoder"
version = "0.1.0"
description = "Transformer-based sentence/passage embeddings as a service."
authors = ["Danilo Miranda Gusicuma <danilo.m.gusicuma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = ">=0.78.0,<0.79.0"
typer = "~0.4.2"
torch = "~2.0.1"
pydantic = "~1.10.5"
uvicorn = "~0.20.0"
python-dotenv = "~0.17.1"
httpx = "0.23.3"
numpy = "1.22.4"
transformers = "~4.26.1"
packaging = "~23.0"
tqdm = "~4.65.0"
mlrun = "~1.2.1"
setuptools = "~67.2.0"
sentence-transformers = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
